UX-614 Convert button utils to ts
